+++ processor/src/test/fixtures/input/com/example/injector/includes/injector/MyOtherInjectorModel.java
@@ -2,7 +2,7 @@ package com.example.injector.includes.injector;
 
 import sting.Injector;
 
-@Injector( includes = MyFragment.class )
+@Injector( injectable = true, includes = MyFragment.class )
 public interface MyOtherInjectorModel
 {
   MyModel getMyModel();